Split DatabaseConfig and DatasourceConfig + move ACL config to security module
